Output nothing if there is no report
With `putStrLn`, even if the result of `showViolations` is empty, the
output is `\n`, resulting on output even if the report is empty.

This is highly unconvenient when `krank` is called on a batch of files,
such as `for i in $(git ls-files); do krank $i; done`

`putStr` does not happen the trailing newline and let this honor to showViolations.

This closes #15.