feat: CLI now handles many files at once
The feature is as simple as parsing `many` `FILE` arguments and mapping
over the list.

However, it introduces a regression because any file which fails during
parsing (file not found, not readable, invalide byte sequence, ...) will
kill the krank process.

I handle that with a `catchAnyDeep` which ensures that no synchroneous
exception will kill the process or stay silent.

In the future, it may be wise to do handle differently some of the
exceptions.

This closes #21. Krank can now be called such as `krank $(git
ls-files)`.

One drawback is that exceptions in a file kills the process