Add threaded execution of file parsing AND queries
Benchmark. 1543 files. 574404 lines. 77 issues

Note that timing for requests are not stable and depends on current
network connectivity.

- pure grep run: 0.029s

- sequential (before this commit):
  --dry-run: 0.650s
  with requests: 39.3s

(memory peak usage was 3Mo)

- with ghc -threaded (this commit). -N1
  --dry-run: 0.750s
  with requestes: 1.708s

- with ghc -threaded (this commit). -N4 (on a 4 core - 8 hyperthread
machine)
  --dry-run: 0.406s
  with requests: 1.108s

(memory usage peak is 24Mo)

The number of thread does not really impact the timming. Actually,
increasing the number of thread may reduce the performances. Most of the
gain is due to the parallel wait for the queries, which can be achieved
with only one thread.